+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * [BREAKING] Incremented MSRV to 1.89.
+* Added WASM bindings for the `Address` type from the miden_objects crate([#1244](https://github.com/0xMiden/miden-client/pull/1244))
 
 ## 0.11.1 (2025-08-31)